Refactored JIX.
JavaImpQuickFix (JIX) now uses javac to compile the current file.  In
the event that there are missing class symbols according to javac's
output, it will attempt a JIS for each one.  There are a number of
advantages to this implementation:

 - Cursor location is maintained.
 - An ant build window is now longer required before running JIX.
 - Only javac is required in the PATH.
 - Error Format and Make Prg are maintained.

This issue mostly resolves #5, but there are still additional work
items required.