Adding Bottom Imports.
The plugin can now be configured to sort imports matching a provided
list of regular expressions to the bottom of the import list (before
static imports if configured to sort statics last).

Updated test cases to include a case for Bottom Imports.  Also, the
tests were updated to not assume default values for the same set of four
plugin options.